+++ feincms3/cleanse.py
@@ -6,26 +6,23 @@ to be, and a strict white-list based HTML sanitizer is the best answer
 I have.
 """
 
+from django.conf import settings
+
 from ckeditor.fields import RichTextField
-from feincms_cleanse import Cleanse
+from html_sanitizer import Sanitizer
+
 
 __all__ = ('CleansedRichTextField', 'cleanse_html')
 
 
-# Site-wide patch of cleanse settings
-# -----------------------------------
-
-Cleanse.allowed_tags['a'] += ('id', 'name')  # Allow anchors
-Cleanse.allowed_tags['hr'] = ()  # Allow horizontal rules
-Cleanse.allowed_tags['h1'] = ()  # Allow H1
-Cleanse.empty_tags += ('hr', 'a')  # Allow empty <hr/> and anchor targets
+sanitizer = Sanitizer(getattr(settings, 'FEINCMS3_SANITIZER_SETTINGS', {}))
 
 
 def cleanse_html(html):
     """
     Pass ugly HTML, get nice HTML back.
     """
-    return Cleanse().cleanse(html)
+    return sanitizer.sanitize(html)
 
 
 class CleansedRichTextField(RichTextField):
@@ -37,18 +34,41 @@ class CleansedRichTextField(RichTextField):
             'default': {
                 'toolbar': 'Custom',
                 'format_tags': 'h1;h2;h3;p;pre',
-                'toolbar_Custom': [
-                    ['Format', 'RemoveFormat'],
-                    ['Bold', 'Italic', '-',
-                     'NumberedList', 'BulletedList', '-',
-                     'Anchor', 'Link', 'Unlink', '-',
-                     'Source'],
-                ],
+                'toolbar_Custom': [[
+                    'Format', 'RemoveFormat', '-',
+                    'Bold', 'Italic', 'Subscript', 'Superscript', '-',
+                    'NumberedList', 'BulletedList', '-',
+                    'Anchor', 'Link', 'Unlink', '-',
+                    'HorizontalRule', 'SpecialChar', '-',
+                    'Source',
+                ]],
             },
         }
 
         # Settings for feincms3.plugins.richtext.RichText
         CKEDITOR_CONFIGS['richtext-plugin'] = CKEDITOR_CONFIGS['default']
+
+    The corresponding ``FEINCMS3_SANITIZER_SETTINGS`` configuration for
+    html-sanitizer_ would look as follows:
+
+        SANITIZER_SETTINGS = {
+            'tags': {
+                'a', 'h1', 'h2', 'h3', 'strong', 'em', 'p', 'ul', 'ol', 'li',
+                'br', 'sub', 'sup', 'hr',
+            },
+            'attributes': {
+                'a': ('href', 'name', 'target', 'title', 'id'),
+            },
+            'empty': {'hr', 'a', 'br'},
+            'separate': {'a', 'p', 'li'},
+            # 'add_nofollow': False,
+            # 'autolink': False,
+            # 'element_filters': [],
+            # 'sanitize_href': sanitize_href,
+        }
+
+    At the time of writing those are the defaults of html-sanitizer_, so you
+    don't have to do anything.
 
     If you want or require a different cleansing function, simply override
     the default with ``CleansedRichTextField(cleanse=your_function)``. The